Add files via upload
This project addresses a big Healthcare issue here in Nigeria among women which is Cervical cancer.  Titled ‘Evaluation of Deep learning models in the classification of Cervical cancer’.  Using Convolutional Neural Network to pass images through a neural network to process images and produce an output and solve a classification problem. Further enhancements will be applied and shared.